backend/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
openai>=1.0.0,<2.0.0
python-dotenv>=1.0.0,<2.0.0
python-multipart>=0.0.9
PyMuPDF>=1.24.0
python-pptx>=0.6.23
python-docx>=1.1.0
firebase-admin>=6.6.0
google-cloud-firestore>=2.20.0
